requirements.txt
# Zynetrax v3.0 — requirements.txt
# Install: pip install -r requirements.txt

# Core
dash==2.17.1
dash-table==5.0.0
plotly==5.22.0
pandas==2.1.3
numpy==1.26.4
openpyxl==3.1.2

# AI / LLM
openai==1.30.0
python-dotenv==1.0.0

# Data processing
duckdb==0.9.2
polars==0.19.19
scikit-learn==1.3.2
statsmodels==0.14.0

# File handling
pdfplumber==0.10.3
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.3.0
opencv-python==4.8.1.78

# Web/API
requests==2.31.0
beautifulsoup4==4.12.2

# Social / Scheduling
apscheduler==3.10.4
cryptography==41.0.7

# Media
gtts==2.5.1
pydub==0.25.1
speechrecognition==3.10.0

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Image generation (optional — large download ~3GB)
# Uncomment to enable Stable Diffusion:
# diffusers==0.24.0
# transformers==4.36.0
# torch==2.3.0 --index-url https://download.pytorch.org/whl/cpu
# accelerate==0.30.0

# Video generation (requires FFmpeg system package)
# moviepy==1.0.3
# ffmpeg-python==0.2.0

# System packages (install separately):
# FFmpeg:   https://ffmpeg.org/download.html
# Tesseract: https://github.com/tesseract-ocr/tesseract
